Refactored test's suites for better readability
